src/requirements.txt
aws-lambda-typing==2.20.0
mysql-connector-python==9.3.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pgvector==0.4.1
pip-review==1.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
